requirements.txt
numpy
pillow
tqdm
matplotlib
torch==1.8.0
torchvision==0.9.0
